requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
Flask==3.0.3
Flask-SocketIO==5.3.7
frozenlist==1.4.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.1.0
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
Pygments==2.18.0
pytest==8.3.3
pytest-cov==5.0.0
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
rich==13.8.1
simple-websocket==1.0.0
textual==0.79.1
tqdm==4.66.5
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.11.1
